pyproject.toml
[project]
name = "pdf-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "boto3>=1.40.57",
    "easyocr>=1.7.2",
    "faiss-cpu>=1.12.0",
    "ipython>=8.37.0",
    "openai>=1.24.0,<2.0.0",
    "qdrant-client>=1.15.1",
    "opencv-python==4.8.1.78",
    "paddleocr>=2.7.0",
    "paddlepaddle>=3.2.0",
    "pdf2image>=1.17.0",
    "pqdm>=0.2.0",
    "pymupdf>=1.26.5",
    "pytesseract>=0.3.13",
    "tabula>=1.0.5",
    "tabula-py>=2.10.0",
    "python-dotenv>=1.1.1",
    "pdfplumber>=0.11.7",
    "pydantic>=2.0.0",
    "tenacity>=8.2.0",
    "tqdm>=4.66.0",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "kaleido>=1.2.0",
    "plotly>=6.4.0",
]
